BT: Add BLE GAIA support.
CRs-Fixed: 2914856

+++ stack/smp/p_256_ecc_pp.h
@@ -26,8 +26,9 @@
 
 #include <stdbool.h>
 #include "p_256_multprecision.h"
+#include "smp_int.h"
 
-typedef unsigned long  DWORD;
+typedef unsigned int DWORD;
 
 typedef struct {
     DWORD x[KEY_LENGTH_DWORDS_P256];
